requirements.txt
pycryptodome==3.19.0
pyautogui==0.9.54
Pillow==10.0.0